minispartan6: Cleanup and modernize.
